Camera et target/texture

+++ include/SuperOpenGL/Face.hpp
@@ -9,17 +9,17 @@ namespace SuperOpenGL {
 class Face
 {
 public:
-  size_t v0, v1, v2;
+unsigned int v0, v1, v2;
 
-  Face();
-  Face(const Face&);
-  Face(const size_t v0, const size_t v1, const size_t v2);
-  ~Face() = default;
+Face();
+Face(const Face&);
+Face(const unsigned int v0, const unsigned int v1, const unsigned int v2);
+~Face() = default;
 
-  Face& operator=(const Face&);
+Face& operator=(const Face&);
 
-  friend std::ostream& operator<<(std::ostream&, const Face&);
-  friend std::istream& operator>>(std::istream&, Face&);
+friend std::ostream& operator<<(std::ostream&, const Face&);
+friend std::istream& operator>>(std::istream&, Face&);
 };
 
 }
